+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Upgrade Sass gem dependency to 3.2.10
 * Fix bug where `ZeroUnit` linter would not report lints for properties with
   lists of values or Sass script
+* Fix bug where `HexFormat` linter would report lints for ID selectors with
+  names that could be hexadecimal color values
 
 ## 0.9.0
 * Add `--show-linters` flag for listing all linters available to scss-lint